+++ priv/repo/migrations/20200630123819_add_provider_to_user_and_make_password_mandatory.exs
@@ -1,0 +1,17 @@
+defmodule Mobilizon.Storage.Repo.Migrations.AddProviderToUserAndMakePasswordMandatory do
+  use Ecto.Migration
+
+  def up do
+    alter table(:users) do
+      add(:provider, :string, null: true)
+      modify(:password_hash, :string, null: true)
+    end
+  end
+
+  def down do
+    alter table(:users) do
+      remove(:provider)
+      modify(:password_hash, :string, null: false)
+    end
+  end
+end