+++ test/mobilizon/users/users_test.exs
@@ -72,14 +72,6 @@ defmodule Mobilizon.UsersTest do
 
     @email "email@domain.tld"
     @password "password"
-    test "authenticate/1 checks the user's password" do
-      {:ok, %User{} = user} = Users.register(%{email: @email, password: @password})
-
-      assert {:ok, _} = Users.authenticate(%{user: user, password: @password})
-
-      assert {:error, :unauthorized} ==
-               Users.authenticate(%{user: user, password: "bad password"})
-    end
 
     test "get_user_by_email/1 finds an user by its email" do
       {:ok, %User{email: email} = user} = Users.register(%{email: @email, password: @password})